+++ src/lib/dglib/include/DgBase.h
@@ -34,7 +34,10 @@
 using namespace std;
 
 #define DGDEBUG         0
-#define DGGRID_VERSION  "7.07" 
+#define DGGRID_VERSION  "7.1" 
+
+// adapted from stackoverflow user Pierre
+#define WHERE fprintf(stderr,"[LOG]%s:%s#%d\n",__PRETTY_FUNCTION__,__FILE__,__LINE__);
 
 ////////////////////////////////////////////////////////////////////////////////
 class DgBase {