+++ src/lib/dglib/include/DgBoundedRF.h
@@ -140,7 +140,10 @@ template<class A, class B, class DB> class DgBoundedRF
                                              
    protected:
 
-      void setFirstAdd (const A& firstAddIn) { firstAdd_ = firstAddIn; }
+      void setFirstAdd (const A& firstAddIn) 
+         { firstAdd_ = firstAddIn;
+           discRF().forceAddress(&(DgBoundedRFBase0::first_), firstAddIn); }
+
       void setLastAdd  (const A& lastAddIn)  { lastAdd_ = lastAddIn; }
       void setEndAdd   (const A& endAddIn)   { endAdd_ = endAddIn; }