+++ src/lib/dglib/include/DgRF.h
@@ -36,6 +36,7 @@ using namespace std;
 
 class DgConverterBase;
 class DgLocation;
+class NuCell;
 
 ////////////////////////////////////////////////////////////////////////////////
 template <class A, class D> class DgRF : public DgRFBase {
@@ -55,6 +56,11 @@ template <class A, class D> class DgRF : public DgRFBase {
       virtual DgLocation* createLocation (const DgLocation& loc,
                                           bool convert = false) const;
 
+// USE_NUCELL is set in MakeIncludes
+#ifdef USE_NUCELL
+      virtual NuCell* createCell (const NuCell& cell, bool convert = false) const;
+#endif
+
       virtual DgLocation* makeLocation (const A& addIn) const;
 
       virtual DgDistanceBase* distance (const DgLocation& loc1, 
@@ -62,6 +68,9 @@ template <class A, class D> class DgRF : public DgRFBase {
                                         bool convert = false) const;
 
       const A* getAddress (const DgLocation& loc) const;
+
+      // deletes old address
+      void forceAddress (DgLocation* loc, const A& addIn) const;
 
       void buildLocVector (DgLocVector* vec, vector<A*> inVec) const;