+++ src/lib/dglib/lib/DgDmdIDGG.cpp
@@ -18,10 +18,9 @@
 *******************************************************************************/
 ////////////////////////////////////////////////////////////////////////////////
 //
-// DgHexIDGG.cpp: DgHexIDGG class implementation
+// DgDmdIDGG.cpp: DgDmdIDGG class implementation
 //
-// Version 7.0 - Kevin Sahr, 11/15/14
-// Version 6.1 - Kevin Sahr, 5/23/13
+// Kevin Sahr, 8/12/20
 //
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -29,141 +28,100 @@
 #include <climits>
 #include <cfloat>
 
-#include "DgHexIDGG.h"
-#include "DgHexIDGGS.h"
-#include "DgHexGrid2DS.h"
+#include "DgDmdIDGG.h"
+#include "DgIDGGS4D.h"
+#include "DgDmdD4Grid2D.h"
+#include "DgDmdD8Grid2D.h"
+#include "DgDmdD4Grid2DS.h"
+#include "DgDmdD8Grid2DS.h"
 #include "DgSeriesConverter.h"
 #include "DgRadixString.h"
 #include "DgBoundedIDGG.h"
 
 ////////////////////////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////
-DgHexIDGG::DgHexIDGG (const DgHexIDGGS& dggs, unsigned int aperture,
-              int res, const string& name, unsigned int precision)
+DgDmdIDGG::DgDmdIDGG (const DgIDGGS4D& dggs, unsigned int aperture,
+              int res, const string& name, bool isD4, unsigned int precision)
    : DgIDGGBase (&dggs, dggs.geoRF(), aperture, res, name, precision),
-	   scaleFac_ (1.0L), rotRads_ (0.0L)
+	   scaleFac_ (1.0L), isD4_ (isD4)
 { 
+//cout << "isD4_: " << ((isD4_) ? "true" : "false") << endl;
    initialize();
 
-} // DgHexIDGG::DgHexIDGG
+} // DgDmdIDGG::DgDmdIDGG
 
 ////////////////////////////////////////////////////////////////////////////////
-DgHexIDGG::DgHexIDGG (const DgHexIDGG& rfIn)
+DgDmdIDGG::DgDmdIDGG (const DgDmdIDGG& rfIn)
    : DgIDGGBase (rfIn.dggs(), rfIn.geoRF(), rfIn.aperture(), 
                  rfIn.res(), rfIn.name(), rfIn.precision()),
-	scaleFac_ (rfIn.scaleFac()), rotRads_ (rfIn.rotRads())
+	scaleFac_ (rfIn.scaleFac()),
+	isD4_ (rfIn.isD4())
 { 
    initialize();
 
-} // DgHexIDGG::DgHexIDGG
+} // DgDmdIDGG::DgDmdIDGG
 
 ////////////////////////////////////////////////////////////////////////////////
-DgHexIDGG::~DgHexIDGG (void) { }
+DgDmdIDGG::~DgDmdIDGG (void) { }
 
 ////////////////////////////////////////////////////////////////////////////////
-const DgHexIDGGS& 
-DgHexIDGG::hexDggs (void) const 
-{ return *(static_cast<const DgHexIDGGS*>(dggs())); }
+const DgIDGGS4D& 
+DgDmdIDGG::dmdDggs (void) const 
+{ return *(static_cast<const DgIDGGS4D*>(dggs())); }
 
 ////////////////////////////////////////////////////////////////////////////////
 void
-DgHexIDGG::initialize (void)
+DgDmdIDGG::initialize (void)
 {
    // verify parameter validity
 
-   string apErrStr = string("DgHexIDGG::initialize(): invalid aperture " + 
+   string apErrStr = string("DgDmdIDGG::initialize(): invalid aperture " + 
                          dgg::util::to_string(aperture()) + 
                          string(" for grid topo ") + gridTopo());
 
-   if (gridTopo() != "HEXAGON")
-   {
-      report("DgHexIDGG::initialize(): invalid grid topo " + gridTopo(), 
+   if (!DgIDGG::isDiamondTopo(gridTopo()))
+      report("DgDmdIDGG::initialize(): invalid grid topo " + gridTopo(), 
              DgBase::Fatal);
 
-      if (aperture() != 3 && aperture() != 4 && aperture() != 7) 
-         report(apErrStr, DgBase::Fatal);
-   }
-
+   if (aperture() != 4) report(apErrStr, DgBase::Fatal);
+ 
    // create some internal data structures
    undefLoc_ = makeLocation(undefAddress());
    sphIcosa_ = new DgSphIcosa(vert0(), azDegs());
 
-   radix_ = (int) sqrt((long double) aperture());
-
-   isAligned_ = true;
+   isAligned_ = false;
    isCongruent_ = false;
 
    // initialize parent values as if this is grid res 0
-   int parentIsClassI = 1;
-   int parentIsClassIII = 0;
    long double parentScaleFac = 1.0;
    unsigned long long int parentNCells = 1;
 
    // get actual parent values if there is a parent grid
-   if (res() > 0) 
-   {
-      const DgHexIDGG& parentIDGG = hexDggs().hexIdgg(res() - 1);
+   if (res() > 0) {
+      const DgDmdIDGG& parentIDGG = dmdDggs().dmdIdgg(res() - 1);
 
-      parentIsClassI = parentIDGG.isClassI();
-      parentIsClassIII = parentIDGG.isClassIII();
       parentScaleFac = parentIDGG.scaleFac();
       parentNCells = parentIDGG.gridStats().nCells();
    }
 
-   // set our values
-
-   if (aperture() == 3)
-   {
-      isClassI_ = !parentIsClassI;
-      isClassIII_ = parentIsClassIII;
-   }
-   else if (aperture() == 4)
-   {
-      isClassI_ = parentIsClassI;
-      isClassIII_ = parentIsClassIII;
-   }
-   else if (aperture() == 7)
-   {
-      isClassI_ = parentIsClassI;
-      isClassIII_ = !parentIsClassIII;
-   }
-
-   allocRes_ = res();
-   if (!isClassI() || isClassIII()) 
-      ++allocRes_;
-
-   // set-up local network to scale so that quad (and consequently tri) edge 
+   // set-up local network to scale so that quad (and consequently dmd) edge 
    // length is 1.0
    ccFrame_ = new DgContCartRF(locNet_, name() + "CC1");
-   grid2DS_ = new DgHexGrid2DS(locNet_, ccFrame(), allocRes() + 1, hexDggs().apSeq(),
-        name() + string("H2DS"));
+//cout << "isD4_: " << ((isD4_) ? "true" : "false") << endl;
+   if (isD4_)
+      grid2DS_ = new DgDmdD4Grid2DS(locNet_, ccFrame(), res() + 1, 4, true, false, name() + string("D4H2DS"));
+   else // D8
+      grid2DS_ = new DgDmdD8Grid2DS(locNet_, ccFrame(), res() + 1, 4, true, false, name() + string("D8H2DS"));
    //cout << "== NEW GRID2DS:" << endl;
    //cout << *grid2DS_;
 
    if (res() == 0)
       maxD_ = 0;
-   else
-   {
-      double factor = parentScaleFac;
-      if (aperture() == 3)
-         factor *= M_SQRT3;
-      else if (aperture() == 4)
-         factor *= 2.0L;
-      else if (aperture() == 7)
-         factor *= M_SQRT7;
+   else {
+      double factor = parentScaleFac * 2.0L; // aperture 4
 
       scaleFac_ = factor;
-      maxD_ = 1.0L;
-
-      // find maxD, which may be for a Class I substrate grid
-      //long long int maxD2 = 0;
-      if (!isClassI())
-         factor *= M_SQRT3;
-
-      if (isClassIII())
-         factor *= M_SQRT7;
-
-      maxD_ = factor - 1.0;
+      maxD_ = factor - 1.0L;
 
       //cout << res() << " " << aperture();
       //cout << " f: " << factor << " maxD: " << maxD_ << endl;
@@ -216,7 +174,7 @@ DgHexIDGG::initialize (void)
                      asinl(sqrt(gridStats_.cellAreaKM() / M_PI) / 
                      (2.0L * DgGeoSphRF::earthRadiusKM())));
 
-} // DgHexIDGG::initialize
+} // DgDmdIDGG::initialize
 
 ////////////////////////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////